# Optimize memory usage of Point sequence (vector) in Way and Area