--- a/libosmscout-binding/src/swig/osmscout/PointSequence.i
+++ b/libosmscout-binding/src/swig/osmscout/PointSequence.i
@@ -1,0 +1,7 @@
+%{
+#include <osmscout/PointSequence.h>
+%}
+
+%include <osmscout/util/GeoBox.h>
+%include <osmscout/PointSequence.h>
+