--- a/libosmscout-import/src/osmscout/import/GenLocationIndex.cpp
+++ b/libosmscout-import/src/osmscout/import/GenLocationIndex.cpp
@@ -477,7 +477,7 @@ namespace osmscout {
             if (ring.IsOuterRing()) {
               std::vector<GeoCoord> coords;
 
-              for (const auto& node : ring.nodes) {
+              for (const auto& node : ring.GetNodes()) {
                 coords.push_back(node.GetCoord());
               }
 
@@ -587,7 +587,7 @@ namespace osmscout {
           if (ring.IsOuterRing()) {
             std::vector<GeoCoord> coords;
 
-            for (const auto& node : ring.nodes) {
+            for (const auto& node : ring.GetNodes()) {
               coords.push_back(node.GetCoord());
             }
 
@@ -751,7 +751,7 @@ namespace osmscout {
 
   bool LocationIndexGenerator::AddLocationAreaToRegion(Region& region,
                                                        const Area& area,
-                                                       const std::vector<Point>& nodes,
+                                                       const PointSequence& nodes,
                                                        const std::string& name,
                                                        double minlon,
                                                        double minlat,
@@ -807,7 +807,7 @@ namespace osmscout {
                                                        const RegionIndex& regionIndex)
   {
     if (ring.IsMasterRing() &&
-        ring.nodes.empty()) {
+        ring.GetNodes().empty()) {
       for (const auto& r : area.rings) {
         if (r.IsOuterRing()) {
           GeoBox boundingBox;
@@ -819,7 +819,7 @@ namespace osmscout {
 
           AddLocationAreaToRegion(*region,
                                   area,
-                                  r.nodes,
+                                  r.GetNodes(),
                                   name,
                                   boundingBox.GetMinLon(),
                                   boundingBox.GetMinLat(),
@@ -838,7 +838,7 @@ namespace osmscout {
 
       AddLocationAreaToRegion(*region,
                               area,
-                              ring.nodes,
+                              ring.GetNodes(),
                               name,
                               boundingBox.GetMinLon(),
                               boundingBox.GetMinLat(),
@@ -926,7 +926,7 @@ namespace osmscout {
           !(minlat>childRegion->maxlat)) {
         // Check if one point is in the area
         for (size_t i=0; i<childRegion->areas.size(); i++) {
-          bool match=IsAreaAtLeastPartlyInArea(way.nodes,childRegion->areas[i]);
+          bool match=IsAreaAtLeastPartlyInArea(way.GetNodes(),childRegion->areas[i]);
 
           if (match) {
             bool completeMatch=AddLocationWayToRegion(*childRegion,way,name,minlon,minlat,maxlon,maxlat);
@@ -945,7 +945,7 @@ namespace osmscout {
     region.locations[name].objects.push_back(ObjectFileRef(way.GetFileOffset(),refWay));
 
     for (size_t i=0; i<region.areas.size(); i++) {
-      if (IsAreaCompletelyInArea(way.nodes,region.areas[i])) {
+      if (IsAreaCompletelyInArea(way.GetNodes(),region.areas[i])) {
         return true;
       }
     }