--- a/libosmscout-import/src/osmscout/import/GenOptimizeAreasLowZoom.cpp
+++ b/libosmscout-import/src/osmscout/import/GenOptimizeAreasLowZoom.cpp
@@ -233,7 +233,7 @@ namespace osmscout
             }
           }
         }
-        newRings.back().SetNodes(new VectorPointSequence(newNodes));
+        newRings.back().SetNodes(newNodes);
 
         r++;
       }