--- a/libosmscout/include/osmscout/Point.h
+++ b/libosmscout/include/osmscout/Point.h
@@ -92,7 +92,7 @@ namespace osmscout {
      */
     Id GetId() const;
 
-    inline const GeoCoord& GetCoord() const
+    inline const GeoCoord GetCoord() const
     {
       return coord;
     }