… Tests/src/CoordinateEncoding.cpp
… libosmscout-import/src/osmscout/import/GenRouteDat.cpp
… libosmscout-import/src/osmscout/import/SortWayDat.cpp
… libosmscout-map/include/osmscout/MapPainter.h
… libosmscout-map/src/osmscout/MapPainter.cpp
… libosmscout/CMakeLists.txt
… libosmscout/include/osmscout/Area.h
… libosmscout/include/osmscout/PointSequence.h
… libosmscout/include/osmscout/RoutingService.h
… libosmscout/include/osmscout/Way.h
… libosmscout/include/osmscout/util/FileWriter.h
… libosmscout/include/osmscout/util/Geometry.h
… libosmscout/include/osmscout/util/Transformation.h
… libosmscout/src/Makefile.am
… libosmscout/src/osmscout/Area.cpp
… libosmscout/src/osmscout/LocationService.cpp
… libosmscout/src/osmscout/PointSequence.cpp
… libosmscout/src/osmscout/RoutePostprocessor.cpp
… libosmscout/src/osmscout/RoutingService.cpp
… libosmscout/src/osmscout/Way.cpp
… libosmscout/src/osmscout/util/FileScanner.cpp
… libosmscout/src/osmscout/util/FileWriter.cpp
… libosmscout/src/osmscout/util/Geometry.cpp
… libosmscout/src/osmscout/util/Transformation.cpp