Hi, my target device (Jolla phone) has limited amount of memory (1 GiB). When some application on foreground occupy more that ~300 MiB, kernels "lowmemorykiller" start killing background apps. Sometimes when foreground app occupy ~500 MiB and system starts email synchronization for example, even foreground application is convicted to die. For this reasons, I want to keep OSMScout (heap) memory usage < 300 MiB. Remaining available space can be used for various caches then.

My experiments with memory profiler shows that *std::vector<Point>* in Way and Area can occupy significant amount of memory:
https://cloud.githubusercontent.com/assets/309458/15384023/bbd950be-1d97-11e6-9bc5-6b5803b4f073.png

But these Point sequences just contains data read from memory mapped database files. So my idea is to create new type *PointSequence* and use it instead of std::vector<Point>. PointSequence can contains just a pointer to memory mapped file and everything else can be evaluated lazy, with minimum of memory allocations. See initial implementation for more details. What do you think about it?

This concept should be good for sequential access but can be slower with random reads. But most parts of library already using iterator-like access to PointSequence. My initial memory measurements looks great. Code itself needs to be cleaned up little bit, I think that it is not ready for merging. But I will welcome every feedback and help.