@Framstag 2016-05-26
I'm not yet convinced of this patch, though some parts of the changes could/should get integrated.

* I do see memory consumption as crucial. So if the library consumes too much memory this has to be fixed.
* In your measurements (if I read the image correctly) the most memory is hold by optimized areas via a list. For me this looks like you are just read a rather huge amount of optimized areas - which of course will result in a huge amount of vector instances. This is a little bit strange, since there should be only a few optimized areas and these should only have a few nodes. Did you analyse this before? Some simple test shows, that I access up to 4-5000 optimized areas (and 2-3000 ways) for a rendering. Are you in the same range? Bigger? It would more sense, to first analyse if the amount of areas could not be reduced (are all 4-5000 areas actually visible, should be not drop to small areas?). It would also make sense to check how much node each area has and if a more aggressive reduction might make more sense. Low level optimizations are OK but only if algorithmic optimizations are already at their limit.
* Memory mapped files require that the sum of the size of all memory mapped files can be mapped into the OS virtual memory. For 32 bit systems this is 4GB. The importer crashes for the same reasons under Linux if process memory reaches around 3,7GB. I do not know how many memory mapped files you can actual open (and hold open at the same time) on iOS or Android. You I would not rely on memory mapped files being available.
* Memory mapped files would be of excellent use of if file representation and memory representation do match. In this case memory mapped files make thing definitely easier. This is not the case for libosmscout because I did the design decision that I trade smaller on disk file size and platform independent format in favour of data reading size (we could reduce the CPU demand and would be able to load bigger data blocks into memory if we would do same changes to the file data model). This could be discussed but I assume that resulting file size would increase by factors. As far as I can see, you fall back to normal vectors if the file is not memory mapped.
* You assume that the file is always open. I fear you will get very strange (catchable?) errors, if the file has been close in the meanwhile?
* What is the impact on the import time? This will also give some hint how high are the cost by trading low memory footprint in favour to CPU an disk IO. Did the rendering on the mobile device take longer?
* You are also moving much data in to the the PointSequence, which again results in some memory footprint.
* I have no problem with dropping standard containers and actually have though of this in the past - if the custom container behaves like a standard one. For example range based loop could be used.
* I also did see the problem with allocation many objects. For me the potential problem however was more the number of (small) objects and the resulting memory overhead. So the question would be: How much would be gain, if we just reduce the number of allocated objects? How could be reduce the number of objects? The solution would be to even more store data in tiles/cells/blocks and always load these in whole. This way I could use one big buffer for all coordinate "strips" in the tile and just address this by offset and count.
* Your solution is more expensive (CPU, IO and overall time wise), if points are visited more than once, because it has to be "loaded" every time again. Currently data is cached in the MapService. So the renderer scan data again and again. Rendering should get more expensive for your solution this way. I plan to move more caching into the renderer. I'm not sure if as a result the render must visit (all) data again after initially processing a tile. On the other side the renderer will definitely have a different data layout then actual raw data, it would not benefit from the PointSequence.

As you see, the patch would be a bigger change and I'm not yet sure if it would move things is the right direction. It definitely make some parts of the code more difficult. It would help if you give me some feedback to better judge, since I do not want to through away you changes light-heartly.

In a result, I plan to analyse the number of optimized areas to see if this can get reduced. This would solve the problem much more elegantly.