@Karry 2016-06-13
Hi Tim. 

I updated rest of modules to work with `PointSequence` container. Here are updated results with fixed [memory leak](https://github.com/Framstag/libosmscout/commit/dfc5cd919608eb4d63d47c9d8dae23c79f66806b):
 - Import process was not slowdown by PointSequence usage 
    - Czech republic import tooks 5 hours, 26 minutes with master, 5:24 with PointSequence branch
 - `PointSequence` has better results with `PerformanceTest`, see [scout-perf.zip](https://github.com/Framstag/libosmscout/files/311304/scout-perf.zip) logs. For example for zoom level 16:
    - avg. memory usage 164 MB -> 126 MB (after loading tile data)
    - avg. DB load 860 ms -> 411 ms
    - avg. Map Rednering 575 ms -> 314 ms

I don't have `PerformanceTest` results from my phone yet, because It don't contains some Qt components like `QDesktopWidget`. It require addition work to compile it.